Added transIR for PRINT (stack strings), PRINT (heap strings) and READ (stored in the heap)
We decided we would store the strings from the get_line function in the heap. This way, we only use exactly as much memory as we need. Naturally, this doesn't mean it's devoid of flaws: we spend twice as much time processing the strings during the read and print (heap) functions. Also, we can't find an easy way to create the concatenation function without using up extra space. We wanted to find a way to point a string to another once they are concatenated but this would mean at least two of the following: it would be very difficult for us to distinguish the first string from the result string, and how would we make MIPS connect one string to another? Likely, we would have to do this during the code generation. For these reasons, we decided to simply use extra space for the concatenations. We can't have it all with a deadline.